+++ packages/block-library/src/audio/save.js
@@ -1,0 +1,15 @@
+/**
+ * WordPress dependencies
+ */
+import { RichText } from '@wordpress/block-editor';
+
+export default function save( { attributes } ) {
+	const { autoplay, caption, loop, preload, src } = attributes;
+
+	return (
+		<figure>
+			<audio controls="controls" src={ src } autoPlay={ autoplay } loop={ loop } preload={ preload } />
+			{ ! RichText.isEmpty( caption ) && <RichText.Content tagName="figcaption" value={ caption } /> }
+		</figure>
+	);
+}
