+++ packages/block-library/src/audio/transforms.js
@@ -1,0 +1,59 @@
+/**
+ * WordPress dependencies
+ */
+import { createBlobURL } from '@wordpress/blob';
+import { createBlock } from '@wordpress/blocks';
+
+const transforms = {
+	from: [
+		{
+			type: 'files',
+			isMatch( files ) {
+				return files.length === 1 && files[ 0 ].type.indexOf( 'audio/' ) === 0;
+			},
+			transform( files ) {
+				const file = files[ 0 ];
+				// We don't need to upload the media directly here
+				// It's already done as part of the `componentDidMount`
+				// in the audio block
+				const block = createBlock( 'core/audio', {
+					src: createBlobURL( file ),
+				} );
+
+				return block;
+			},
+		},
+		{
+			type: 'shortcode',
+			tag: 'audio',
+			attributes: {
+				src: {
+					type: 'string',
+					shortcode: ( { named: { src } } ) => {
+						return src;
+					},
+				},
+				loop: {
+					type: 'string',
+					shortcode: ( { named: { loop } } ) => {
+						return loop;
+					},
+				},
+				autoplay: {
+					type: 'srting',
+					shortcode: ( { named: { autoplay } } ) => {
+						return autoplay;
+					},
+				},
+				preload: {
+					type: 'string',
+					shortcode: ( { named: { preload } } ) => {
+						return preload;
+					},
+				},
+			},
+		},
+	],
+};
+
+export default transforms;
