+++ packages/block-library/src/block/index.js
@@ -12,16 +12,12 @@ export const name = 'core/block';
 
 export const settings = {
 	title: __( 'Reusable Block' ),
-
 	category: 'reusable',
-
 	description: __( 'Create content, and save it for you and other contributors to reuse across your site. Update the block, and the changes apply everywhere it’s used.' ),
-
 	supports: {
 		customClassName: false,
 		html: false,
 		inserter: false,
 	},
-
 	edit,
 };
