+++ packages/block-library/src/calendar/index.js
@@ -12,18 +12,12 @@ export const name = 'core/calendar';
 
 export const settings = {
 	title: __( 'Calendar' ),
-
 	description: __( 'A calendar of your site’s posts.' ),
-
 	icon: 'calendar',
-
 	category: 'widgets',
-
 	keywords: [ __( 'posts' ), __( 'archive' ) ],
-
 	supports: {
 		align: true,
 	},
-
 	edit,
 };
