+++ packages/block-library/src/categories/index.js
@@ -13,17 +13,12 @@ export const name = 'core/categories';
 
 export const settings = {
 	title: __( 'Categories' ),
-
 	description: __( 'Display a list of all categories.' ),
-
 	icon,
-
 	category: 'widgets',
-
 	supports: {
 		align: true,
 		html: false,
 	},
-
 	edit,
 };
