+++ packages/block-library/src/classic/index.js
@@ -1,7 +1,6 @@
 /**
  * WordPress dependencies
  */
-import { RawHTML } from '@wordpress/element';
 import { __, _x } from '@wordpress/i18n';
 
 /**
@@ -10,6 +9,7 @@ import { __, _x } from '@wordpress/i18n';
 import edit from './edit';
 import icon from './icon';
 import metadata from './block.json';
+import save from './save';
 
 const { name } = metadata;
 
@@ -17,11 +17,8 @@ export { metadata, name };
 
 export const settings = {
 	title: _x( 'Classic', 'block title' ),
-
 	description: __( 'Use the classic WordPress editor.' ),
-
 	icon,
-
 	supports: {
 		className: false,
 		customClassName: false,
@@ -29,12 +26,6 @@ export const settings = {
 		// confusing UX, because of its similarity to the 'Convert to Blocks' button.
 		reusable: false,
 	},
-
 	edit,
-
-	save( { attributes } ) {
-		const { content } = attributes;
-
-		return <RawHTML>{ content }</RawHTML>;
-	},
+	save,
 };
