+++ packages/block-library/src/classic/save.js
@@ -1,0 +1,10 @@
+/**
+ * WordPress dependencies
+ */
+import { RawHTML } from '@wordpress/element';
+
+export default function save( { attributes } ) {
+	const { content } = attributes;
+
+	return <RawHTML>{ content }</RawHTML>;
+}
