+++ packages/block-library/src/code/index.js
@@ -2,7 +2,6 @@
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
-import { createBlock } from '@wordpress/blocks';
 
 /**
  * Internal dependencies
@@ -10,6 +9,8 @@ import { createBlock } from '@wordpress/blocks';
 import edit from './edit';
 import icon from './icon';
 import metadata from './block.json';
+import save from './save';
+import transforms from './transforms';
 
 const { name } = metadata;
 
@@ -17,47 +18,12 @@ export { metadata, name };
 
 export const settings = {
 	title: __( 'Code' ),
-
 	description: __( 'Display code snippets that respect your spacing and tabs.' ),
-
 	icon,
-
 	supports: {
 		html: false,
 	},
-
-	transforms: {
-		from: [
-			{
-				type: 'enter',
-				regExp: /^```$/,
-				transform: () => createBlock( 'core/code' ),
-			},
-			{
-				type: 'raw',
-				isMatch: ( node ) => (
-					node.nodeName === 'PRE' &&
-					node.children.length === 1 &&
-					node.firstChild.nodeName === 'CODE'
-				),
-				schema: {
-					pre: {
-						children: {
-							code: {
-								children: {
-									'#text': {},
-								},
-							},
-						},
-					},
-				},
-			},
-		],
-	},
-
+	transforms,
 	edit,
-
-	save( { attributes } ) {
-		return <pre><code>{ attributes.content }</code></pre>;
-	},
+	save,
 };
