+++ packages/block-library/src/code/save.js
@@ -1,0 +1,3 @@
+export default function save( { attributes } ) {
+	return <pre><code>{ attributes.content }</code></pre>;
+}
