+++ packages/block-library/src/code/transforms.js
@@ -1,0 +1,35 @@
+/**
+ * WordPress dependencies
+ */
+import { createBlock } from '@wordpress/blocks';
+
+const transforms = {
+	from: [
+		{
+			type: 'enter',
+			regExp: /^```$/,
+			transform: () => createBlock( 'core/code' ),
+		},
+		{
+			type: 'raw',
+			isMatch: ( node ) => (
+				node.nodeName === 'PRE' &&
+				node.children.length === 1 &&
+				node.firstChild.nodeName === 'CODE'
+			),
+			schema: {
+				pre: {
+					children: {
+						code: {
+							children: {
+								'#text': {},
+							},
+						},
+					},
+				},
+			},
+		},
+	],
+};
+
+export default transforms;
