+++ packages/block-library/src/column/save.js
@@ -1,0 +1,22 @@
+/**
+ * External dependencies
+ */
+import classnames from 'classnames';
+
+/**
+ * WordPress dependencies
+ */
+import { InnerBlocks } from '@wordpress/block-editor';
+
+export default function save( { attributes } ) {
+	const { verticalAlignment } = attributes;
+	const wrapperClasses = classnames( {
+		[ `is-vertically-aligned-${ verticalAlignment }` ]: verticalAlignment,
+	} );
+
+	return (
+		<div className={ wrapperClasses }>
+			<InnerBlocks.Content />
+		</div>
+	);
+}
