+++ packages/block-library/src/columns/index.js
@@ -1,16 +1,7 @@
 /**
- * External dependencies
- */
-import classnames from 'classnames';
-
-/**
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
-
-import {
-	InnerBlocks,
-} from '@wordpress/block-editor';
 
 /**
  * Internal dependencies
@@ -19,6 +10,7 @@ import deprecated from './deprecated';
 import edit from './edit';
 import icon from './icon';
 import metadata from './block.json';
+import save from './save';
 
 const { name } = metadata;
 
@@ -26,32 +18,14 @@ export { metadata, name };
 
 export const settings = {
 	title: __( 'Columns' ),
-
 	icon,
-
 	description: __( 'Add a block that displays content in multiple columns, then add whatever content blocks you’d like.' ),
-
 	supports: {
 		align: [ 'wide', 'full' ],
 		html: false,
 	},
-
 	deprecated,
-
 	edit,
-
-	save( { attributes } ) {
-		const { columns, verticalAlignment } = attributes;
-
-		const wrapperClasses = classnames( `has-${ columns }-columns`, {
-			[ `are-vertically-aligned-${ verticalAlignment }` ]: verticalAlignment,
-		} );
-
-		return (
-			<div className={ wrapperClasses }>
-				<InnerBlocks.Content />
-			</div>
-		);
-	},
+	save,
 };
